refactor: move PlanStep to domain to fix layer boundary violation
